backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# Data Processing
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
statsmodels==0.14.0
networkx==3.2.1

# HTTP and WebSocket
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# Financial Data
ccxt==4.1.77

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Environment
python-dotenv==1.0.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator==2.2.0
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
